gen_id(): Polish help-file description
The description is more explicit in what it does. Before, the
description referred to "Name" and "Sector" but this was as bit
confusing because there was no column -- or value of a column --
called "Name" or "Sector". Not it is more clear.

However, I am not convinced it's the best approach. As it is now,
the description may expose the implementation details  too much.