New replace_customer_name()
Closes 2DegreesInvesting/r2dii.match#8
Closes  maurolepore/r2dii.match#1

FIX
* Fix R CMD check
* Fix white space issue

POLISH
* Spellcheck
* Style
* Add and polish tests and helpfile and examples
* Evgeny is now an author & contributor
* Remove excesive infrastructure for internal data
    *  data-raw/ and sysdata.rds
* Use packages: tibble, rlang, r2dii.utils, purrr
* Remove redundant @author

FEATURES
* Names replacements can now have columns from and to in
  lower/upper case.
* Use argument defaults via NULL

REFACTOR
* DRY
* Extract functions and varibles for clarity
* Slide variable closer to where it is used
* Remove unnecessary variable and return()
* Rename arguments.
* Name replacements & ownership are now internal (see [1] below)

[1] This follows a conversation with @jdhoffa:

> for now we should hardcode them in, and perhaps later we can add
> a feature, allowing a user to incorporate their own reduction
> rules.
>
> -- https://github.com/maurolepore/r2dii.match/pull/1#issuecomment-544523541

The idea is to make this code flexible and easy to maintain.

This code relies on replacement rules that come from a
spreadsheet-like dataset. For contributors, it may be easiest to
edit that dataset from a .csv file, but such a change involves many
steps (e.g. changes in data-raw/ in data/) that create friction for
maintainers. So my approach here is to brind the data close to
the fucntion that uses it, and ask contributors to simply edit
a tribble, by adding a new line directly from the GitHub interface.
This maybe unfamiliar but is easy enough that I would like to
try it.

Also, this approach reduces the chances of introducing bugs due
to white space. White space is important in this feature, and
it is easy to lose or convert to NA the white space because
of the defaults of readr::read_csv() (trim_ws = TRUE, na = c('')).
A tribble makes white space explicit and easy to see.

Thanks @jdhoffa
Signed-off-by: Mauro Lepore <maurolepore@gmail.com>