Test for printed output only where needed
* Point to the PR discussing why this test exists.
* Remove this rare test from elsewhere.

Details:
This commit clarifies that in id_by_loantaker_sector() I tested
the output is printed to the console because we had detected a
but by which the result  of the function was assigned to an object
but not returned at the end. It is best practice to write a test
once a bug is fixed, so that the test does not come back in the
future. But this is function-specific, so we don't need to write
this relatively rare test for other functions. Thus I removed it
from sector_bridge(), which has no record of such a bug.