Refer to arguments by name
It's ok to call the first 1 or two data-arguments by position
but a third details-argument is probabbly better to name
explicitly. Here the first argument is piped, the second is matched
by position but it's kind of obvious by the name `*from_ald()`, and
the third is now named explicitely. We should be good.