Coerce level cols as characters
Closes #146

If level columns are not of type character, we now coerce them to
character. Compared to throwing an error this seems more reasonable
as it is clear that these columns should be characters, and any
type can be coerced to character -- hence we can do it
programatically on behalf of the user.

Thanks @jdhoffa