Shorten intro.
I agree that the intro was dense, but the best solution I think
is not to split it but to say the same more succinctly. I try my
best for the last sentence of the fisrt paragraph to be the one
that introduces the hole article -- this is a particularly rigid
expectation that readers inconciously have:

> The opening paragraph of a document is
  unique. Readers have remarkably rigid
  expectations of what to find where. They
  almost always expect that the point will
  be articulated in the final sentence.
  -- https://georgegopen.net/wp-content/uploads/2019/12/Five-Varieties-of-Point-Placement-The-Construction-of-the-English-Paragraph-Part-V.pdf

You can go over more than one paragraph and move the point to the
last sentence of the last pargraph, but then you need to signal
this 'unexpected' structure by separating the intro from the rest
of the document -- for example with extra white-space, or with a
line of asterisks like this:

****************

But this seems an overkill here becuase the split would now leave
only two sentences in the first paragraph and only 1 sentence in
the second and last paragraph of the introduction.