library(dplyr, warn.conflicts = FALSE)
This makes the test output a bit more quiet so it's easier to
read. It does increase the risk of conflicts but I think only by
a bit, as the verbose output leads to ignoring it. A better way
to avoid conflicts would be with the package conflicted. But I
think dplyr functions are well konwn so the chances of we adding
a conflict by accident are low, so maybe it's not worthwhile
adding the dependency on conlicted?