Avoid error on R CMD check with R <= 3.4
It's best practice to use on.exit() with add = TRUE and
after = FALSE; or instead to use withr::defer(). But add = TRUE
was added after 3.4, and I want to avoid a dependency on withr.
So I just go with on.exit() and add = TRUE, as this only really
matters when you call on.exit() more than once.