Simpler NEWS so we have less things to maintain
The News menu has a "Releases" section that I now find confusing
because there are more releases than entries under News/Releases,
pointing to external articles. But we don't write articles for
most small patches.

This PR removes the headings "Releases" and "Changelog", to
instead expose the changelog directly under the News navbar.

--

Alternativelly, we may rename the heading from "Releases" to
"Announcements" (which is less confusing), and/or add a single
link to the blog.